requirements.txt
streamlit
pandas
numpy
folium
streamlit_folium
sklearn.ensemble
